[libc] Add attribute for extern replacing macros
Globals like `stdout` or `errno` are in fact macros to unspecified
globals of the standard libc. In Frama-C's libc we define our own
globals as targets for those macros. This commit adds an attribute to
the global so that the macro that it intends to replace can be retrieved
by a plugin.