[Eva] Evaluation engine: fixes a soundness bug when the oracle produces alarms.
During an evaluation, a domain can use the oracle to request the evaluation
of another unrelated expression. If this secondary evaluation produces any
alarm, the resulting values may have been reduced accordingly, but the alarms
will not be emitted (as they might be unrelated to the current statement).
In this case, the values resulting from the secondary evaluation are not sound
over-approximations, and should not be used by the domains.

When an evaluation requested by an oracle leads to any alarm, do not record its
results in the cache, which is also used as the result of the primary evaluation.

This bug should not happen with the current abstract domains, as they only
request new evaluations of expressions already evaluated on previous statement,
so the produced alarms should have already been emitted.
However, it could happen in the near future, if domains start using the oracle
on new expressions.