[e-acsl] e-acsl-gcc.sh: add concurrency flags automatically
By grepping its output we can use Frama-C's pthread detection to add
concurrency flags to the invocation of gcc.
We have to remove -q and -verbose 0 from the Frama-C flags in wrapper.sh
in order to obtain the pthread detection message.
We no longer need explicit concurrency flags in the test files, so we
are remove them.