[Eva] In offsetmaps, fixes function [update_imprecise_everywhere].
[update_imprecise_everywhere validity origin value offsm] does no longer remove
nodes of [offsm] which are not valid according to [validity].

The merge of two offsetmap is only defined for offsetmaps of the same size,
so all offsetmaps representing the same base must always have the same size.
But for dynamically allocated memory, the size of weak bases can change during
the analysis.
To account for possible future reallocation, offsetmaps representing dynamically
allocated memory are created with a size equal to the maximum possible size
when reallocating memory (see Base.max_allocable), with all non-allocated bits
bound to bottom.
This must be preserved by [update_imprecise_everywhere]: ranges outside the
validity must remain unchanged, to not change the offsetmap size. Otherwise,
the merge of offsetmaps representing dynamically allocated memory may crash.