[e-acsl] fix warning about unsupported assigns clauses
Fixes https://git.frama-c.com/pub/frama-c/-/issues/2676

When an assigns clause is encountered the warning message emitted
suggests that the unsupported clause is inside a behaviour:
Warning: E-ACSL construct `assigns clause in behavior' is not yet supported.
This occurs whether the assigns clause is inside of a behaviour or not.

Assign clauses are currently not supported no matter whether they are
inside of behaviours or not. Therefore it suffices to simply remove the
`in behavior' part from the message.