[e-acsl] Extract typedef from struct decl on recursive type
The type `eacsl_assert_data_value_t` is recursive with a member `next`
that is a pointer to `eacsl_assert_data_value_t`. In the previous
version the pointer was directly to the struct instead of the typedef'd
type, but the typedef'd type adds an attribute.

The commit puts the typedef before the declaration of the struct so that
the typedef version `eacsl_assert_data_value_t` with the attribute can
be used everywhere.