[e-acsl] memoise generated functions by signature
Hitherto function were memoised by profile.
This prevented sharing of the generated code in cases of two or more
function calls to the same function, with the calls having different
profiles but all profiles mapping to the same type.

Now only one function is generated for such a set of calls.