[kernel] In Parameter_sig, removes keys from Value_datatype.
Value_datatype is the signature required to build map parameters.
The related key was given as argument to the function converting a string into
a value for the key, allowing for a different conversion according to the key.

This is not the case anymore, which simplifies the different implementations of
Value_datatype.