[Eva] fix a crash when perf stack and call stack were desynchronized
Eva_perf stop operation is only valid if the stack is not empty. `Eva_utils` assumes the call stack and the `Eva_perf` stack are in sync and call `Eva_perf.stop` only when the call stack is not empty.

However, the call stack and perf stack were not always synchronized: `post_analysis` cleared the call stack but not the perf stack. This is often not a problem since the call stack is usually empty at the end of an analysis... except when there is an error that stops the analysis.

This commits splits callstack resets to `Eva_utils.reset_call_stack` and `Eva_utils.clear_call_stack` depending on whether stats must be preserved (end of analysis) or not (start of analysis). `Eva_perf.reset` call is moved to to `Eva_utils.reset_call_stack` thus every `Eva_utils` operations on callstack are synchronized with `Eva_perf`.