[e-acsl] inductives: fix mode selection
Foreign predicates CAN be used in complete mode. The mode selection
during the extraction has to match the preceding mode selection of the
mode analysis, so the filter was not correct.

A problem which persists for now is the fact that predicate applications
inside a logic_info definition are not handled by the Logic_normalizer.
The reason is that visitors do not descend into logic_info definitions.
For the example at hand this results in a failure to use the extracted
version of Q inside of P.

[e-acsl] foreign_complete.c:14: Warning:
  E-ACSL construct
  `inductive definitions outside of subset described in the reference manual'
  is not yet supported.
  Ignoring annotation.