[e-acsl] Ghost bases are declared static in the runtime library
In the E-ACSL runtime library, ghost globals are used as bases for newly
allocated contracts and assert data so that Eva can check the
specifications even though it does not support allocates directives yet.

The globals were declared and defined in header files, which means that
they could have ended up defined in several compilation units, which is
forbidden by the standard. This commit declare the values as `static`
so that the variables do not escape their compilation units.