[kernel] Clarify Filepath.(/) documentation
The `Filepath.(/)` operator is equivalent to `Filepath.concat` which
concatenate a directory and a file. However the documentation of the
operator was using the example `Filepath.(file / ext)` which make it
seem like an extension is appended to a filename. This commit replaces
it with `Filepath.(dir / file)` like the example for `Filepath.concat`.