[e-acsl] integrate use of Error.retrieve_preprocessing
It seems that Error.retrieve_preprocessing is used at various places
with a string as an argument that describes the entity being processed.
This leads to really strange error messages like
Failure: quantified predicate was not performed on construct \forall int i;

So first of all we rectify the error by using a string describing the
kind of analysis.

Secondly, we expose only the version of Quantifier.get that already
applies Error.retrieve_preprocessing to its result.