[e-acsl] output information about failing dev tests
Hitherto the error output is redirected into the log file.
This makes it difficult to track down the cause of an error, especially
when it occurs on the CI. This commit ensures that the error message is
logged to the console.

This commit includes some fixes proposed by Basile (@bdesloges).