[e-acsl] Filter "Aborted" message in dev test
In dev tests, if the program expectedly crash, the shell generates an
"Aborted" message that is not captured by standard redirection and might
confuse the user. This commit uses a grouping expression to be able to
capture that message so that it appears instead in the regular .err.log
file. In addition, this message is filtered so that the line number of
wrapper.sh and the PID of the running program do not appear.