[e-acsl] inductives: fuse normalisation into extraction
As formal substitutions are now already generated during the mode
analysis the normalisation was not doing much at all anymore. It only
applied the substitutions and generated some equations for the
conclusion.
It did however not generate equations for recursive hypotheses of the
kind P_fun2(n) ≡ x, which had to be generated during extraction (once
P_fun2 actually has been extracted). So, as such equations needed to be
generated anyway during the extraction it seemed simpler (and for
efficient) to perform everything during the extraction.