[wp] Prefix .why3.conf dependency with file:
The path to the dependency `.why3.conf` is absolute and on windows that
means starting with `C:\` (or another drive letter). Ptests identifies
the kind of dependency by finding the first colon, which means that on
Windows this dependency will be of kind "C". This commit explicitely put
the dependency with the kind `file:` so that the full path is correctly
used by Ptests.