[Eva] In Results API, fixes a crash when a function pointer evaluates to top.
In the evaluation engine, [eval_function] returns `Top (instead of failing)
if [Value.resolve_functions] returns `Top; the failure is done by the
interpretation of call in Transfer_stmt.

This fixes the crash of [Results.eval_callee]: if the analysis has not been
run or results have not been saved, it returns Top (without specific warning).