[alias] add (hardly) pathological example
https://git.frama-c.com/frama-c-indus/merce/-/issues/25

The example does not have any wrong aliases, but it does reveal a
problem in Abstract_state.is_included.
It considers the latter a subgraph of the former:
0:{ p } → 1:{ z }   3:{ d }
0:{ p } → 1:{ d }   2:{ z }

It seems to me that this only occurs in benign cases in the sense that
it doesn't negatively affect aliases. However some users of this plugin
might be interested primarily in the graphs not only as a means to
obtain alias information.

This bug only materialised after recent reworkings of function calls for
optimisation purposes, where now node indices between to-be-merged
graphs have more overlaps.