[alias] Abstract_state.is_included: simplify
The number of successors of a vertex is 0 or 1 in the overwhelming
majority of cases. If not, it is bounded by the number of the fields of
the struct it represents. So the overhead of converting lists into
sequences and maps is surely not justified.

Also: remove an obsolete TODO comment