[e-acsl] no longer generate unsound code for unsupported labels
In the old code the (Not_yet) exceptions encountered while preprocessing
labels were lost, since they were only stored in a hash table that
associated the at_data (containing the exception) with the statement
that the label refers to.

In the case of unsupported labels, such a statement could not be
identified, so the exception was never stored and thus never raised.

Now in Labels we store the at_data (including encountered errors) in all
cases, associating them with the predicate/term the label occurred in.
In Translate_ats, we reuse this store at_data, raising the exception if
one was present.

Fixes https://git.frama-c.com/pub/frama-c/-/issues/2744