[e-acsl] inductives: favour predicate over function extraction
In some edge cases, when a foreign inductive predicate is used, and if
in its result position there is not a variable that needs to be bound,
then instead of extracting a function of P and generating a comparison
of the form
f_P(...) = res
we rather extract an executable predicate and generate
P(..., res)

It is an edge case, since if P is extractable in complete mode, this is
what happens anyway.

The generated code is a little bit nicer, that is all.