[e-acsl] simplify translation for label Here (fixes exception)
The Here-label does not require any preprocessing in order to be
translated, it can simply be ignored.

The proper fix would be to apply the label analysis on new
predicates/functions, but we run into a problem here:
The code in the Labels module presumes the existence of a kernel
function (and some stmt) for which the current term/predicate is
analysed. However in the case of inductive extraction there is no such
kernel function (or statement).

So for this proper fix we would need to render the label analysis
possible for predicates/terms that are not associated with a kernel
function, but this would require a deep change of the
Analyses_types.at_data type. Since labels are to be overhauled soon
anyway, and since logic functions are exclusively supported for the Here
label anyway, we stick to this simple fix.

Fixes https://git.frama-c.com/frama-c/frama-c/-/issues/1604