[e-acsl] inductives: remove quantifier substitution
Since the result of the extraction does not have any quantifiers it does
not matter whether we substitute them or not before extraction.

As in the subsequent commit we are going to substitute logic variables
by terms instead of by logic variables we have to exclude quantifiers
from the substitution (as we cannot put terms into an ∀).