[e-acsl] inductives: substitute logic variables by terms
For now we have been substituting logic variables by other logic
variables. In the next commit we are going to allow deriving a
substitution  x ↦ n - 1  for a conclusion P(n + 1).
In other words, we are going to permit deriving substitutions for terms
by allowing for terms of specific forms function inversion.

Co-authored-by: Basile Desloges <basile.desloges@cea.fr>