[e-acsl] fix marking-as-unsound functions/predicates
If a Here-specialised predicate/function is derived from a possibly
unsound predicate/function it has to inherit the unsoundness property.

Some tests now emit warnings of the sort
+count.c:26: Warning: no sound verdict for RTE (guess: ok).
along with pointer values. As they may change between executions we have
to add the option --no-assert-print-data in order to stabilise these tests.