[e-acsl] inductives: emit unsound warning only for negative verdicts
The unsoundness comes from overlaps in the case of function extractions,
e.g. the first constructor yielding false for a proposition while the
second constructor yields true for the same proposition. In that case
the verdict should be true, but the true verdict is masked by the false
verdict.
In any case, there is no unsoundness in the case of true verdicts.

Also, reduce the number of these kinds of warnings, eliminating
repeated ones.